move sorting data into usecase
