+++ SocialCareCaseViewerApi/V1/Gateways/IProcessDataGateway.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using SocialCareCaseViewerApi.V1.Boundary.Requests;
 using SocialCareCaseViewerApi.V1.Boundary.Response;
@@ -15,6 +14,5 @@ namespace SocialCareCaseViewerApi.V1.Gateways
         CareCaseData? GetCaseById(string recordId);
 
         Task<string> InsertCaseNoteDocument(CaseNotesDocument caseNotesDoc);
-        IOrderedEnumerable<CareCaseData> SortData(string sortBy, string orderBy, List<CareCaseData> response);
     }
 }
